Uncompleted project
I failed to complete this project mostly because of the failure to set up the environment for the C++ third party libraries. It was in October 2023 on my laptop, I tried to setup the environment for C++ third party libraries however I failed to do so. So I decided I better implement those projects in other programming languages like JavaScript, Python, C# and Java.  So
I set out to learning the following programming languages that allows me easy to setup environment : C#, Java, JavaScript and Dart, TypeScript, Python, Go